+++ lint-staged.js
@@ -1,8 +1,3 @@
 module.exports = {
-  '*.{js,jsx,ts,tsx}': [
-    'eslint --max-warnings=0'
-    // TODO- Add it once ready with tests
-    // "npm test",
-  ],
-  '*.{js,jsx,ts,tsx,css}': ['prettier --write', 'eslint --fix']
+  '*.{js,jsx,ts,tsx,css}': ['prettier --write', 'eslint --fix', 'yarn test']
 };