+++ src/components/movieCard.tsx
@@ -1,10 +1,12 @@
 import React, { memo, useCallback, useMemo } from 'react';
-import { Image, StyleSheet, TouchableOpacity } from 'react-native';
+import { Image, StyleSheet, TouchableOpacity, View } from 'react-native';
+import { MaterialIcons } from '@expo/vector-icons';
 
 import { screenNames } from 'constants/screenNames';
 import { navigationService } from 'navigations/navigationService';
 import { useStore } from 'store/movies';
 
+import { colors } from 'styles/colors';
 import containers from 'styles/containers';
 
 interface MovieCardProps {
@@ -15,11 +17,19 @@ const styles = StyleSheet.create({
   poster: {
     width: '100%',
     height: '100%'
+  },
+  favoriteContainer: {
+    position: 'absolute',
+    zIndex: 1,
+    top: 8,
+    right: 8
   }
 });
 
 const MovieCard: React.FC<MovieCardProps> = ({ id }) => {
-  const movieData = useStore()?.selectMovieById(id);
+  const { selectMovieById, selectIsFavoriteMovie } = useStore();
+  const movieData = selectMovieById(id);
+  const isFavorite = selectIsFavoriteMovie(id);
 
   const handlePress = useCallback(() => {
     navigationService.navigate(
@@ -41,6 +51,11 @@ const MovieCard: React.FC<MovieCardProps> = ({ id }) => {
   return (
     <TouchableOpacity style={containers.flex1} onPress={handlePress}>
       <Image style={styles.poster} source={source} />
+      <View style={styles.favoriteContainer}>
+        {isFavorite ? (
+          <MaterialIcons name="favorite" size={24} color={colors.rose} />
+        ) : null}
+      </View>
     </TouchableOpacity>
   );
 };