+++ src/styles/colors.ts
@@ -1,4 +1,5 @@
 export const colors = {
   black: '#000000',
-  white: '#ffffff'
+  white: '#ffffff',
+  rose: '#e11d48'
 };